+++ portals/admin/src/main/webapp/source/src/app/components/Governance/ComplianceDashboard/APICompliance/Compliance.jsx
@@ -95,7 +95,7 @@ export default function Compliance(props) {
                 width='full'
                 title={(
                     <FormattedMessage
-                        id='Governance.Overview.Compliance.title'
+                        id='Governance.ComplianceDashboard.Compliance.title'
                         defaultMessage='Compliance Summary - {artifactName}'
                         values={{ artifactName }}
                     />
@@ -104,15 +104,15 @@ export default function Compliance(props) {
             >
                 <Box sx={{ display: 'flex', alignItems: 'center', paddingBottom: 4 }}>
                     <RouterLink
-                        to='/governance/overview'
+                        to='/governance/compliance'
                         style={{
                             display: 'flex', alignItems: 'center', textDecoration: 'none', color: 'inherit',
                         }}
                     >
                         <ArrowBackIcon />
                         <FormattedMessage
-                            id='Governance.Overview.Compliance.back.to.overview'
-                            defaultMessage='Back to Overview'
+                            id='Governance.ComplianceDashboard.Compliance.back.to.compliance'
+                            defaultMessage='Back to Compliance Dashboard'
                         />
                     </RouterLink>
                 </Box>
@@ -140,7 +140,7 @@ export default function Compliance(props) {
                         sx={{ fontWeight: 'medium' }}
                     >
                         <FormattedMessage
-                            id='Governance.Overview.Compliance.check.progress'
+                            id='Governance.ComplianceDashboard.Compliance.check.progress'
                             defaultMessage='Compliance Check in Progress'
                         />
                     </Typography>
@@ -150,7 +150,7 @@ export default function Compliance(props) {
                         align='center'
                     >
                         <FormattedMessage
-                            id='Governance.Overview.Compliance.check.progress.message'
+                            id='Governance.ComplianceDashboard.Compliance.check.progress.message'
                             defaultMessage='The compliance check is currently in progress. This may take a few moments.'
                         />
                     </Typography>
@@ -164,7 +164,7 @@ export default function Compliance(props) {
             width='full'
             title={(
                 <FormattedMessage
-                    id='Governance.Overview.Compliance.title'
+                    id='Governance.ComplianceDashboard.Compliance.title'
                     defaultMessage='Compliance Summary - {artifactName}'
                     values={{ artifactName }}
                 />
@@ -179,20 +179,20 @@ export default function Compliance(props) {
             }}
             >
                 <RouterLink
-                    to='/governance/overview'
+                    to='/governance/compliance'
                     style={{
                         display: 'flex', alignItems: 'center', textDecoration: 'none', color: 'inherit',
                     }}
                 >
                     <ArrowBackIcon />
                     <FormattedMessage
-                        id='Governance.Overview.Compliance.back.to.overview'
-                        defaultMessage='Back to Overview'
+                        id='Governance.ComplianceDashboard.Compliance.back.to.compliance'
+                        defaultMessage='Back to Compliance Dashboard'
                     />
                 </RouterLink>
                 <Typography variant='body2'>
                     <FormattedMessage
-                        id='Governance.Overview.Compliance.api.owner'
+                        id='Governance.ComplianceDashboard.Compliance.api.owner'
                         defaultMessage='API Owner: {owner}'
                         values={{ owner: artifactOwner }}
                     />
@@ -225,7 +225,7 @@ export default function Compliance(props) {
                                 sx={{ fontWeight: 'bold', mb: 2 }}
                             >
                                 <FormattedMessage
-                                    id='Governance.Overview.Compliance.policy.adherence.summary'
+                                    id='Governance.ComplianceDashboard.Compliance.policy.adherence.summary'
                                     defaultMessage='Policy Adherence Summary'
                                 />
                             </Typography>
@@ -243,7 +243,7 @@ export default function Compliance(props) {
                                 sx={{ fontWeight: 'bold', mb: 2 }}
                             >
                                 <FormattedMessage
-                                    id='Governance.Overview.Compliance.ruleset.adherence'
+                                    id='Governance.ComplianceDashboard.Compliance.ruleset.adherence'
                                     defaultMessage='Ruleset Adherence'
                                 />
                             </Typography>
@@ -254,7 +254,7 @@ export default function Compliance(props) {
                                         id: 0,
                                         value: statusCounts.passed,
                                         label: `${intl.formatMessage({
-                                            id: 'Governance.Overview.Compliance.passed',
+                                            id: 'Governance.ComplianceDashboard.Compliance.passed',
                                             defaultMessage: 'Passed',
                                         })} (${statusCounts.passed})`,
                                     },
@@ -262,7 +262,7 @@ export default function Compliance(props) {
                                         id: 1,
                                         value: statusCounts.failed,
                                         label: `${intl.formatMessage({
-                                            id: 'Governance.Overview.Compliance.failed',
+                                            id: 'Governance.ComplianceDashboard.Compliance.failed',
                                             defaultMessage: 'Failed',
                                         })} (${statusCounts.failed})`,
                                     },
@@ -295,7 +295,7 @@ export default function Compliance(props) {
                                 sx={{ fontWeight: 'bold', mb: 2 }}
                             >
                                 <FormattedMessage
-                                    id='Governance.Overview.Compliance.ruleset.adherence.summary'
+                                    id='Governance.ComplianceDashboard.Compliance.ruleset.adherence.summary'
                                     defaultMessage='Ruleset Adherence Summary'
                                 />
                             </Typography>
