+++ portals/admin/src/main/webapp/source/src/app/components/Governance/ComplianceDashboard/PolicyAdherenceTable.jsx
@@ -18,16 +18,16 @@
  */
 
 import React from 'react';
-import { Box, Chip, Typography, Tooltip, LinearProgress } from '@mui/material';
+import { Box, Chip, Typography, TableRow, TableCell, LinearProgress } from '@mui/material';
 import ListBase from 'AppComponents/AdminPages/Addons/ListBase';
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { Link as RouterLink } from 'react-router-dom';
-import ErrorIcon from '@mui/icons-material/Error';
-import WarningIcon from '@mui/icons-material/Warning';
-import InfoIcon from '@mui/icons-material/Info';
+import Stack from '@mui/material/Stack';
+import CheckCircleIcon from '@mui/icons-material/CheckCircle';
+import CancelIcon from '@mui/icons-material/Cancel';
 import GovernanceAPI from 'AppData/GovernanceAPI';
 import { useIntl } from 'react-intl';
-import ApiIcon from '@mui/icons-material/Api';
+import PolicyIcon from '@mui/icons-material/Policy';
 import Utils from 'AppData/Utils';
 
 /**
@@ -37,37 +37,46 @@ import Utils from 'AppData/Utils';
 function apiCall() {
     const restApi = new GovernanceAPI();
     return restApi
-        .getComplianceStatusListOfAPIs()
+        .getPolicyAdherenceForAllPolicies()
         .then((result) => {
-            return result.body.list;
+            const policies = result.body.list;
+
+            // Fetch policy adherence details for each policy
+            // TODO: optimize
+            return Promise.all(
+                policies.map(async (policy) => {
+                    try {
+                        const adherenceDetails = await restApi.getPolicyAdherenceByPolicyId(policy.id);
+                        return {
+                            ...policy,
+                            evaluatedArtifacts: adherenceDetails.body.evaluatedArtifacts || []
+                        };
+                    } catch (error) {
+                        console.error(`Error fetching adherence for policy ${policy.id}:`, error);
+                        return {
+                            ...policy,
+                            evaluatedArtifacts: []
+                        };
+                    }
+                })
+            );
         })
         .catch((error) => {
             throw error;
         });
 }
 
-
-export default function ApiComplianceTable() {
+export default function PolicyAdherenceTable() {
     const intl = useIntl();
 
+    // TODO: reuse this function in other components
     const renderProgress = (followed, total, status) => {
-        if (status === 'PENDING') {
+        if (status === 'UNAPPLIED') {
             return (
                 <Typography variant="body2" color="textSecondary">
                     {intl.formatMessage({
-                        id: 'Governance.Overview.APICompliance.pending',
-                        defaultMessage: 'N/A - Waiting for policy evaluation',
-                    })}
-                </Typography>
-            );
-        }
-
-        if (status === 'NOT_APPLICABLE') {
-            return (
-                <Typography variant="body2" color="textSecondary">
-                    {intl.formatMessage({
-                        id: 'Governance.Overview.APICompliance.no.policies',
-                        defaultMessage: 'N/A - No policies to evaluate',
+                        id: 'Governance.ComplianceDashboard.PolicyAdherence.no.apis',
+                        defaultMessage: 'N/A - No APIs to evaluate',
                     })}
                 </Typography>
             );
@@ -81,8 +90,8 @@ export default function ApiComplianceTable() {
                 <Box sx={{ display: 'flex', mb: 0.5 }}>
                     <Typography variant="body2" sx={{ fontWeight: 'bold' }} color="textSecondary">
                         {intl.formatMessage({
-                            id: 'Governance.Overview.APICompliance.followed.count',
-                            defaultMessage: '{followed}/{total} Followed',
+                            id: 'Governance.ComplianceDashboard.PolicyAdherence.compliant.count',
+                            defaultMessage: '{followed}/{total} Compliant',
                         }, { followed, total })}
                     </Typography>
                 </Box>
@@ -103,74 +112,35 @@ export default function ApiComplianceTable() {
         );
     };
 
-    const renderComplianceIcons = (severityBasedRuleViolationSummary) => {
-        // get the error, warn, info counts
-        let errorCount = 0;
-        let warnCount = 0;
-        let infoCount = 0;
-
-        severityBasedRuleViolationSummary.forEach((severity) => {
-            if (severity.severity === 'ERROR') {
-                errorCount = severity.violatedRulesCount;
-            } else if (severity.severity === 'WARN') {
-                warnCount = severity.violatedRulesCount;
-            } else if (severity.severity === 'INFO') {
-                infoCount = severity.violatedRulesCount;
-            }
-        });
-
-        return (
-            <Tooltip title={`Error: ${errorCount || 0}, Warn: ${warnCount || 0}, Info: ${infoCount || 0}`}>
-                <Box sx={{ display: 'flex', gap: 1 }}>
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <ErrorIcon color="error" sx={{ fontSize: 16 }} />
-                        <Typography variant="caption" sx={{ ml: 0.5 }}>
-                            {errorCount || 0}
-                        </Typography>
-                    </Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <WarningIcon color="warning" sx={{ fontSize: 16 }} />
-                        <Typography variant="caption" sx={{ ml: 0.5 }}>
-                            {warnCount || 0}
-                        </Typography>
-                    </Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <InfoIcon color="info" sx={{ fontSize: 16 }} />
-                        <Typography variant="caption" sx={{ ml: 0.5 }}>
-                            {infoCount || 0}
-                        </Typography>
-                    </Box>
-                </Box>
-            </Tooltip>
-        );
-    };
-
-    const columProps = [
+    const policyColumProps = [
         {
             name: 'id',
             options: { display: false }
         },
         {
-            name: 'info',
-            options: { display: false }
-        },
-        {
             name: 'name',
             label: intl.formatMessage({
-                id: 'Governance.Overview.APICompliance.column.api',
-                defaultMessage: 'API',
+                id: 'Governance.ComplianceDashboard.PolicyAdherence.column.policy',
+                defaultMessage: 'Policy',
             }),
             options: {
                 customBodyRender: (value, tableMeta) => (
                     <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <RouterLink to={`/governance/overview/api/${tableMeta.rowData[0]}`} style={{ display: 'flex', alignItems: 'center', textDecoration: 'none', color: 'inherit' }}>
-                            {tableMeta.rowData[1].name}
+                        <RouterLink
+                            to={`/governance/policies/${tableMeta.rowData[0]}`}
+                            style={{ display: 'flex', alignItems: 'center', textDecoration: 'none', color: 'inherit' }}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                        >
+                            {value}
                             <OpenInNewIcon sx={{ ml: 0.5, fontSize: 16 }} />
                         </RouterLink>
                     </Box>
                 ),
                 setCellProps: () => ({
-                    style: { width: '30%' },
+                    style: {
+                        width: '30%'
+                    },
                 }),
                 setCellHeaderProps: () => ({
                     sx: {
@@ -187,22 +157,20 @@ export default function ApiComplianceTable() {
         {
             name: 'status',
             label: intl.formatMessage({
-                id: 'Governance.Overview.APICompliance.column.status',
+                id: 'Governance.ComplianceDashboard.PolicyAdherence.column.status',
                 defaultMessage: 'Status',
             }),
             options: {
                 customBodyRender: (value) => (
                     <Chip
-                        label={Utils.mapComplianceStateToLabel(value)}
-                        color={value === 'COMPLIANT' ? 'success' :
-                            value === 'NON_COMPLIANT' ? 'error' :
-                                value === 'PENDING' ? 'warning' : 'default'}
+                        label={Utils.mapPolicyAdherenceStateToLabel(value)}
+                        color={value === 'FOLLOWED' ? 'success' : value === 'VIOLATED' ? 'error' : 'default'}
                         size="small"
                         variant="outlined"
                     />
                 ),
                 setCellProps: () => ({
-                    style: { width: '20%' },
+                    sx: { width: '20%' },
                 }),
                 setCellHeaderProps: () => ({
                     sx: {
@@ -217,49 +185,25 @@ export default function ApiComplianceTable() {
             },
         },
         {
-            name: 'policyAdherenceSummary',
+            name: 'artifactComplianceSummary',
             options: { display: false }
         },
         {
-            name: 'severityBasedRuleViolationSummary',
+            name: 'evaluatedArtifacts',
             options: { display: false }
         },
         {
-            name: 'policies',
+            name: 'progress',
             label: intl.formatMessage({
-                id: 'Governance.Overview.APICompliance.column.policies',
-                defaultMessage: 'Policies',
+                id: 'Governance.ComplianceDashboard.PolicyAdherence.column.apis',
+                defaultMessage: 'APIs',
             }),
             options: {
                 customBodyRender: (value, tableMeta) => {
-                    const followed = tableMeta.rowData[4]?.followed || 0;
-                    const violated = tableMeta.rowData[4]?.violated || 0;
-                    const total = followed + violated;
-                    const status = tableMeta.rowData[3];
+                    const followed = tableMeta.rowData[3]?.compliant || 0;
+                    const total = (tableMeta.rowData[3]?.nonCompliant + followed) || 0;
+                    const status = tableMeta.rowData[2];
                     return renderProgress(followed, total, status);
-                },
-                setCellProps: () => ({
-                    style: { width: '40%' },
-                }),
-                setCellHeaderProps: () => ({
-                    sx: {
-                        paddingTop: 0,
-                        paddingBottom: 0,
-                        '& .MuiButton-root': {
-                            fontWeight: 'bold',
-                            fontSize: 'small'
-                        },
-                    },
-                }),
-            },
-        },
-        {
-            name: 'compliance',
-            label: ' ',
-            options: {
-                customBodyRender: (value, tableMeta) => {
-                    const severityBasedRuleViolationSummary = tableMeta.rowData[5] || [];
-                    return renderComplianceIcons(severityBasedRuleViolationSummary);
                 },
                 setCellHeaderProps: () => ({
                     sx: {
@@ -275,6 +219,46 @@ export default function ApiComplianceTable() {
         },
     ];
 
+    const renderExpandableRow = (rowData) => {
+        return (
+            <TableRow>
+                <TableCell colSpan={3} />
+                <TableCell>
+                    <Stack direction="column" spacing={2} sx={{ flexWrap: 'wrap' }}>
+                        {/* TODO: Find a better way to display all those */}
+                        {rowData[4].map((artifact) => (
+                            <Box
+                                key={artifact.id}
+                                sx={{
+                                    display: 'flex',
+                                    alignItems: 'center',
+                                    gap: 1
+                                }}
+                            >
+                                {artifact.status === 'COMPLIANT' ?
+                                    <CheckCircleIcon color="success" sx={{ fontSize: 16 }} /> :
+                                    <CancelIcon color="error" sx={{ fontSize: 16 }} />
+                                }
+                                <RouterLink
+                                    to={`/governance/compliance/api/${artifact.id}`}
+                                    style={{
+                                        display: 'flex',
+                                        alignItems: 'center',
+                                        textDecoration: 'none',
+                                        color: 'inherit'
+                                    }}
+                                >
+                                    {artifact.info.name}
+                                    <OpenInNewIcon sx={{ ml: 0.5, fontSize: 16 }} />
+                                </RouterLink>
+                            </Box>
+                        ))}
+                    </Stack>
+                </TableCell>
+            </TableRow>
+        );
+    };
+
     const emptyStateContent = (
         <Box
             sx={{
@@ -284,7 +268,7 @@ export default function ApiComplianceTable() {
                 padding: 3
             }}
         >
-            <ApiIcon
+            <PolicyIcon
                 sx={{
                     fontSize: 60,
                     color: 'action.disabled',
@@ -298,8 +282,8 @@ export default function ApiComplianceTable() {
                 sx={{ fontWeight: 'medium' }}
             >
                 {intl.formatMessage({
-                    id: 'Governance.Overview.APICompliance.empty.content',
-                    defaultMessage: 'No APIs Available',
+                    id: 'Governance.ComplianceDashboard.PolicyAdherence.empty.content',
+                    defaultMessage: 'No Governance Policies Available',
                 })}
             </Typography>
             <Typography
@@ -308,8 +292,8 @@ export default function ApiComplianceTable() {
                 align="center"
             >
                 {intl.formatMessage({
-                    id: 'Governance.Overview.APICompliance.empty.helper',
-                    defaultMessage: 'Create APIs to start evaluating their compliance.',
+                    id: 'Governance.ComplianceDashboard.PolicyAdherence.empty.helper',
+                    defaultMessage: 'Create a new governance policy to start governing the APIs.',
                 })}
             </Typography>
         </Box>
@@ -317,7 +301,7 @@ export default function ApiComplianceTable() {
 
     return (
         <ListBase
-            columProps={columProps}
+            columProps={policyColumProps}
             apiCall={apiCall}
             searchProps={false}
             emptyBoxProps={{
@@ -326,17 +310,10 @@ export default function ApiComplianceTable() {
             addButtonProps={false}
             showActionColumn={false}
             useContentBase={false}
+            enableCollapsable={true}
+            renderExpandableRow={renderExpandableRow}
             options={{
                 elevation: 0,
-                selectableRows: 'none',
-                customBodyRender: (value) => value,
-                setTableProps: () => ({
-                    style: {
-                        '& .MuiTableCell-root': {
-                            border: 'none'
-                        }
-                    }
-                })
             }}
         />
     );
